Change to float[]s

+++ src/main/java/com/notloc/targettruetile/ImprovedTileIndicatorsUtil.java
@@ -53,9 +53,9 @@ public class ImprovedTileIndicatorsUtil {
         }
 
         int vCount = model.getVerticesCount();
-        int[] x3d = model.getVerticesX();
-        int[] y3d = model.getVerticesY();
-        int[] z3d = model.getVerticesZ();
+        float[] x3d = model.getVerticesX();
+        float[] y3d = model.getVerticesY();
+        float[] z3d = model.getVerticesZ();
 
         int[] x2d = new int[vCount];
         int[] y2d = new int[vCount];
